Use libraryName for deprecation dependency (#122)
Previously, the code incorrectly re-used the `@types` package name for
the deprecation dependency, which is also noted in the readme as the
correct package to install in the future.

This is correct 90%+ percent of the time because `@types/X` usually provides
types for package `X`. Sometimes, however, there is no matching `X`,
so that `@types/X` should tell users to install some other package `Y` when it's deprecated.

For example, `@types/xumm-api` was for a non-npm package that eventually got published
to npm as `xumm-sdk` and at the same
time started shipping types. So the deprecation entry looked like this:

```json
{
  "libraryName": "xumm-sdk",
  "typingsPackageName": "xumm-api",
  "sourceRepoURL": "https://github.com/XRPL-Labs/XUMM-SDK",
  "asOfVersion": "0.1.4"
}
```

There IS no `xumm-api`, only an `xumm-sdk`, so it doesn't make sense to
point people to `xumm-api` when `@types/xumm-api` is deprecated.